requirements.txt
asgiref==3.8.1
colorama==0.4.6
coverage==7.6.4
Django==5.1.2
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.2
drf-yasg==1.21.8
Faker==30.8.1
inflection==0.5.1
iniconfig==2.0.0
marshmallow==3.23.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
